🎨 Task 10: extract shared PlaybackManager jvmTest ProgressTracker helper
